threadpool changes
